Add V2 API for creating new Silence
When users want to create a Silence through Promgen, the current ProxySilences API only
allows passing the name and value of the matcher via the "label" field in the request.
And it cannot pass the isEqual and isRegex values.
We have added a new API, ProxySilencesV2, which allows passing the full set of values:
name, value, isEqual, an isRegex, with a request body similar to when calling the API
of Alertmanager.