Improve filtering active silences on Service and Project Detail pages
Previously, we only supported creating and filtering silences with the '=' operator.
Now that we support all operators, the list of active silences will be displayed on the
Service Detail and Project Detail pages, which have also been adjusted accordingly.

For example: For 'Project P' belonging to 'Service S', the following silences will be
considered active silences related to 'Project P':
- Silence 1: project=P, service=S
- Silence 2: project=~P*, service=~S*
- Silence 3: project!=X, service!=Y
- Silence 4: project!~X*, service!~Y*
- Silence 5: service=S (In Alertmanager, a silence can still apply to an alert even if the
  silence has fewer matchers than the alert).
Previously, only Silence 1 was considered, while the remaining silences were not
counted.